refactor: Cars에서 이동 후 결과 상태를 반환하도록 구현
Cars의 내부 상태를 숨기기 위해(getter 를 public 으로 공개하지 않도록) Cars에서 반환하는 것이 더 좋겠다 판단하여 이렇게 수정하였습니다.
이에 대한 카일의 의견이 궁금해요!
