+++ src/main/java/racingcar/domain/game/RacingCarGame.java
@@ -2,6 +2,7 @@ package racingcar.domain.game;
 
 import racingcar.domain.car.Car;
 import racingcar.domain.car.Cars;
+import racingcar.domain.car.MovedResult;
 
 import java.util.List;
 
@@ -32,20 +33,16 @@ public class RacingCarGame {
     }
 
     /**
-     * Cars.move()의 결과로 MovedResult 를 반환할지,
-     * 혹은 이곳에서 반환할 지 고민중입니다.
-     * <p>
-     * Cars 에서는 자동차들과 관련된 책임을 담당하는 것이 좋다고 생각하는데, 차들의 이동 결과 역시 Cars 에서 담당하는 것이 좋다고 생각합니다.
-     * 그러나 매 Lap 마다 결과를 보여준다는 것은 게임에서 필요로 하는 비즈니스 로직이라 생각하여, 이곳에서 구현하였습니다.
-     * 카일님은 이에 대해 어떻게 생각하시나요?
+     * 조금 더 고민해 본 결과,Cars의 내부 상태를 숨기기 위해(getter 를 public 으로 공개하지 않도록)
+     * Cars에서 반환하는 것이 더 좋겠다 판단하였어요!
+     * 이에 대한 카일의 의견이 궁금해요!
      */
     public MovedResult race() {
         if (!hasMoreLap()) {
             throw new IllegalStateException("이미 모든 바퀴를 진행했습니다.");
         }
-        cars.move(numberGenerator);
         currentLap = currentLap.increase();
-        return new MovedResult(cars);
+        return cars.move(numberGenerator);
     }
 
     public boolean hasMoreLap() {
